+++ app/templates/config/csrf.js
@@ -1,5 +1,0 @@
-/**
- * Cross-Site Request Forgery Protection Settings
- */
-
-module.exports.csrf = false;