+++ app/templates/config/log.js
@@ -7,5 +7,5 @@ module.exports.log = {
      * The order of precedence for log levels from lowest to highest is:
      * silly, verbose, info, debug, warn, error, silent
      */
-    level: 'info'
+    level: 'silly'
 };