+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+Version 0.2.0
+===
+
+- Automatic installing npm dependencies;
+- Add questions for database name, user, password and adapter;
+- Add in connections config all supported adapters;
+- Remove unused middleware from express;
+- Remove unused parts from ```package.json```;
+
 Version 0.1.0
 ===